\name{alphabet-package}
\alias{alphabet-package}
\alias{alphabet}
\docType{package}
\title{
\packageTitle{alphabet}
}
\description{
\packageDescription{alphabet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{alphabet}
\packageIndices{alphabet}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{alphabet}

Maintainer: \packageMaintainer{alphabet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
