\name{AlphabetPackage-package}
\alias{AlphabetPackage-package}
\alias{AlphabetPackage}
\docType{package}
\title{Alphabet
\packageTitle{AlphabetPackage}
}
\description{
\packageDescription{AlphabetPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AlphabetPackage}
\packageIndices{AlphabetPackage}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{AlphabetPackage}

Maintainer: \packageMaintainer{AlphabetPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
